Fix to uninstaller/db clean up/option changes
Updated the uninstaller to be aware of the sub/modules. This way the user when they uninstall a package it will remove the sub/module if it is empty. Cleaned the db up so that they are empty. Changed the print out of the options so that if there is none then it will display no package installed