Update updater.py
Saving changes to the updater. Currently all files except for the cfg and json files will be replaced automatically. This is a good step in the right direction. The cfg and json files will need to be able to be modified and append the data that is needed to the database files so that the user will not lose any of the custom scanners that they have created for the program.