Param update and cfg
Changes made to the cfg file so that they users have to accept the terms. Also changes made to the params so that they can now use the xmrig pcap right out of the box. with the param being set correctly with the location change of the pcap file dir.