Updater changed
Changes made to the updater so that when the user updates the program that some of the files will not be moved until the program is terminating. This allows for us to change the cfg and json files with out worrying about them being overriden on termination. This is a stage change.