Update updater.py
Adding more files that wont be copied over. These files will not be changed a lot and the md files more then likely wont be looked at besides online. Will remove the md later when more work is being done in those files. That is just not the area of focus right now.