Updater
The updater is working for the modules/submodules/packages. There is more work that needs to be done to clean up the output and make sure that it is displaying information relative to the user. But over all this is working well. Also will need to do a package_info updater, this one might take a little while to make.