Update to command line
There has been a massive change to the command line usage. There are a lot more options to navigate around. Examples of valid commands.
Example options:
Module: challenges
Sub: navy
Package: usb

append:
---------
/challenges
/navy/usb
usb

replace:
challenges/navy
challenges/

This is a more stream lined way to get around. There will be more to come with validation of the commands but this is working for the parsing and moving of the pages, not the validation of existence of the module/submodule/package.