Printer and Projects update
There was a few big changes that happened in this update. The printer object is now the primary and the old console output object has been removed. Also output to the console in the effected areas has been changed to using the printer object, except for the report object. Also, the projects option has been re-added to the program, there will be more updates on that to come.