Massive changes
Migrating over to using MongoDB for the server instead of TinyDB. This has involved interfaces being built. This is going to be a slow process. Removed the application configuration and replaced it with the core configuration. Currently the program is checking to see if the Nightcap DB exists in the Mongo Server. Error catches for if the server is down, and started initializer for the server if there is no instance of the Nightcap Database