Updated System Configuration
This is a very good change and fixed many issues that were starting to be developed by unfortunate bad development. The configuration for the system is now a singleton and is being with all of the needed objects correctly. The super calls are fixed and the code change to reflect the removal of parameters