Massive refactoring
There has been some massive changes to the structure of the program. A new plugin has been built, the nightcapcli. This is specifically built for well the cli control. The main application has been cleaned up and basically cleared out. The legal will get moved at some point. Besides moving files and a few new names to the files there is not much of a change.