Simplifying Interfaces
Simplifying the interfaces so that they are more specific and can be built off of easier. Trying to set the code up for a Singleton change from using the decorator to the metaclass which allows for thread safety. There are now two objects for connecting to the mongo server. A connection object that just checks to see if the database is live and another that is an operational connection for crud operations. Objects have been changed to reflect their needed implementation.