Fix to singleton
Removed the decorator singleton and all of the singletons are using the metaclass singleton. This is a great fix, this required a few interface changes and removal of a few to avoid collisions but it worked out in the end. There will be more to come with this change to hopefully enforce the interface itself but right now the interface object is working correctly.