Server start/stop/status
The server start/stop/status is fixed for the config command. This is a great change and needed to be fixed. Also, there is clean up again when the program exits to make sure that the server has been shutdown. The NightcapCoreSimpleServer had to be changed from being a singleton. This was due to the way the configuration was built, but this was for the best. I also changed the banner so that it looks cleaner.