Docker integration + Django site init
So this is a massive commit. The program has been successfully been rolled over to using docker. Haven't tested the ability to connect to a remote container, but locally it is working very well. The containers are being built correctly and things are being handled for errors very well and will adjust the state of the start of the program accordingly to get the user connected. The mongo server when it is created it sets a username/password for the instance. There is also a Django container getting created, there is a base site there that will be used for the reporting site. The console has been rolled over to using the docker containers as well, starting/stopping/etc.