Clean up to the cmds
This is a clean up / staging commit for the cmds. I'am going to be attempting to roll the base usage of the cmds to using an observer pattern. There will be more clean up to come as well. Due to the time constraints on getting this program running I have to put documenting on the side. I will be adding a lot of documentation to the program as soon as the presentation is done.