Observer change
There has been a pretty big change to how the pages are handled for calling the modules/submodules/packages, not from the users point of view but from the code side. Observers are now being used, to control the pages, this allows for dynamic page creation/destruction based on dispatch calls. This is almost completely bug free there is a small issue with some of the paths but it is almost the, committing for a restore point.