Observer, completion change, clean up
The observer is working as good as it can be for right now it is not exactly how I want it to be with removing the parent objects but right now it is working extremely well and is generating all child objects needed.
Ex:
Current Path: root
User wants: testing/navy/usb
All objects will be generated and correctly pushed, this is using an observer to make the changes with a factory to create the end objects with the correct inheritances. I am extremely happy with the outcome of the roll over.

There has also been changes to the tab complete, when the user double tabs the tab for a command to see the options it will then show the attached options. The biggest impact that this had was the ability to use the tab complete with the installed packages for the program. This is working for most commands but right now it is late and I need to commit the code. This was a lot of fine work to get this here. This commit is mainly fine tuning and a little cleaning up from the last few commits as well, there were some bad implementations. This was my first time really working with an observer inside of a cli application, I think it went well.

Will be locking down the use command a little for now. This is just going to be a side effect for a little bit until I have the time to fix the issue, with the parent object removal, I need to build out some scanners and the website.