Clean up of configurations, and package running
This is a big change, the package configuration is inheriting the base config for the program for the packagecli this was a great change. Big change to the way the users run their own custom packages, they will need to added methods to their base object to be able to then utilize the run() call to actually run their program. More to come when documentation is done about how every thing is working. I only have a few more days left until I have to present this project