Update to the scanning
There has been an update to the scanning so that there is now a console print option where the user is allowed to generate a console report to show to the user. This is working well and there is not a need to change the way that this is being implemented. This will allow us to use the Strategy pattern down the road.