Update to the Scanner
When the user is scanning pcap files there was an issue with the data not being passed to the right object to get things setup. This is fixed and the scanner is working right. This also fixes the issue with the scanner not opening the right file when trying to scan the pcap file data. There is also a little clean up that was done for the printing to the console.