CLI fix
There was a big issue with the cli which required a little bit of a rebuild. That is okay tho the issue was caused by a configuration object which has been removed. The dynamic functionality of the console has been limited a little because of it but for now it is working and presentable. This will be fixed when the presentation is over.