Clean up to Initiation of the program
This is a commit for the clean up of the initiation of the program. This makes it easier to see what is going on when the program has started and is needing to be initialized. (Download images, set up dbs, etc.). This is a nice improvement and it is functioning smoothly. To come is more creation stuff for Django.