Getting the Django server ready for mongodb connection
The Django container is getting ready to start connecting to the mongo container and using it for auth, storage, etc. This is a great fix. Currently the connection is working but environment is not ready for the connection. That will be in the next commit. For this to work correctly, we ended up having to change the mongodb port back.