Multi Package Update
Preparing the program to start allowing other package types, currently working creating one called redteam. This is a pretty empty wrapper class that allows the end package developer to create pretty much whatever they want to. So far so good. This also requires an extra parameter to be added to the package_info.json file. The documentation in the wiki will be updated accordingly.