Update to un/installer/signing
There has been a big change to they way that they packages are packed and signed for distribution. When the user signs the package there is a .ncp file that will get created that can then be used to install the package. There was also an issue with the dev settings when the user would enter the settings and enter and exit the dev section more than once the console leading title would add more than one dev option to it. This was an easy fix. There is a little more that needs to be done for the installing, so that it will clean up the tmp location that is being created.