Fixes to the restore process
There was an issue with the restore process, where the imports that were required were not being installed correctly for the packages to be used. This has been fixed and it is working really well now. The backup/restore process is pretty much fixed. There is some more things that will need to be done with cleaning the folders/db for the restore. Also, this will be an option to itself called clean in the backup section.