Backup/Restore/Clean Complete
The backup, restore, and clean features have been completed and are working extremely well. As it stands right now, I feel that this feature is complete as I need to focus on other parts of the project. This does need some more work but nothing extremely major. There is an issue with restoring the projects collection but right now that is not on the top of the list since right now this feature is most likely not going to be used a lot but all of the data IS being saved correctly. A fix for this will come in a few days there are more pressing things I need to get done.