Updater/Future Structure changes
Update to the updater so that it is working with branches. This is a safety commit, there is more that needs to be done for the updater to be completed. The config file needs to be updated. There has also been changes to get ready for network proxy implementations for the whole program. This will allows users to use the socks4/5 proxies or the standard http requests. This has not been implemented up the options in the settings have been prepared. Also, rolled over the updater to a command pattern implementation.