Massive Changes to docker and django abilities
This change consists of big changes to the docker interactions of the program. Docker calls are now using the docker lib and not calling to the command line to do what they were doing. Removed deprecated HTML pages and reenabled the Django container. Updated start of the program to handle the checking and initialization of the program in a more stream lined manner. This is not quite done for stream lining the entry of the program. There are a few more things that need to be done to the docker/mongo/django objects with decoupling and abstraction that needs to be done. Updated requirements for the application as well to include pgpy, docker, six and python-bsonjs. For the start up of the program when initializing, there is a form of password validation and email validation that has been implemented for creation of the main account for the user.