Big changes / Reporting Working
This commit has many different changes inside of it.
The biggest change that this commit has is that the front end
Of the program has been cleaned up and allows users to quickly be able to launch the application.
The code has been streamlined to allow for the users to be able to ‘recover’ the environment
Depending on what stage that it is in, this makes more sense when you look at the start of the program
We can select to options to perform on the current environment or to change the server that we are pointing at.
The reporting object has been created and is able to be used for writing data to the databases for the projects, users
Have the option(s) if they would like to keep/replace the report for a project selected if the package had already been ran before
For the project. This is working well and there will need to be more done on this to make it more useful with actually being able to
Display the data on the site to the users of the program. That will be the next step. The client side has been changed as far as what objects are allowed to be used when the user is creating their own custom packages for the program,  this helps me to limit the scope right now.  Also both the django/mongo helpers have been changed which is based more on inheritance now than what was there before and is more streamlined for the usage of the objects.