Update to Install abilities
Upgraded the program to allow for github installation of programs. This is working but needs a lot more testing. This allows for users to create a 'wrapper' for the github program that they would like to use and install it like a normal package with getting the source code from github. For this to be optimal removal of the reporting/django information needs done; since users will be able to use other programs and not ones that are built specifically for NightCap the reporting will not work correctly as it is right now.